updating code to be human readable and executed by anyone
