--- limits/combine_logger.out
@@ -1,9 +1,0 @@
-INFO: (function: improveOnce) CascadeMinimizer.cc: 157 -- Minimisation configured with Type=Minuit2, Algo=Migrad, strategy=1, tolerance=0.1
-INFO: (function: improveOnce) CascadeMinimizer.cc: 180 -- Minimisation finished with status=0
-INFO: (function: improveOnce) CascadeMinimizer.cc: 185 -- Minimization success! status=0
-INFO: (function: improveOnce) CascadeMinimizer.cc: 157 -- Minimisation configured with Type=Minuit2, Algo=Migrad, strategy=1, tolerance=0.1
-INFO: (function: improveOnce) CascadeMinimizer.cc: 180 -- Minimisation finished with status=0
-INFO: (function: improveOnce) CascadeMinimizer.cc: 185 -- Minimization success! status=0
-INFO: (function: improveOnce) CascadeMinimizer.cc: 157 -- Minimisation configured with Type=Minuit2, Algo=Migrad, strategy=1, tolerance=0.1
-INFO: (function: improveOnce) CascadeMinimizer.cc: 180 -- Minimisation finished with status=0
-INFO: (function: improveOnce) CascadeMinimizer.cc: 185 -- Minimization success! status=0
